requirements.txt
PySide6>=6.6.0
mss>=9.0.1
pyautogui>=0.9.54
Pillow>=10.0.0
# playwright>=1.40.0  # deferred — browser mode not yet implemented
